corrected ex except ex.4

+++ task_1.rb
@@ -3,7 +3,7 @@ puts "Введите Ваше имя"
 name = gets.chomp
 puts "Введите Ваш рост"
 height = gets.chomp
-weight = (height.to_i - 110) * 1.15
+weight = (height.to_f - 110) * 1.15
 if weight > 0 
     puts "#{name}, Ваш идеальный вес #{weight}"
 else
